Some half-implemented refactoring
