+++ packages/gtmcore/gtmcore/dataset/storage/local.py
@@ -1,5 +1,7 @@
+import shutil
+
 from gtmcore.dataset import Dataset
-from gtmcore.dataset.storage.backend import UnmanagedStorageBackend
+from gtmcore.dataset.storage.backend import StorageBackend
 from typing import List, Dict, Callable, Optional
 import os
 
@@ -11,7 +13,7 @@ from gtmcore.dataset.manifest.manifest import Manifest, StatusResult
 logger = LMLogger.get_logger()
 
 
-class LocalFilesystem(UnmanagedStorageBackend):
+class LocalFilesystemBackend(StorageBackend):
 
     def _backend_metadata(self) -> dict:
         """Method to specify Storage Backend metadata for each implementation. This is used to render the UI
@@ -140,7 +142,7 @@ more, check out the docs here: [https://docs.gigantum.com](https://docs.gigantum
         Returns:
             bool
         """
-        return True
+        return False
 
     def update_from_remote(self, dataset, status_update_fn: Callable) -> None:
         """Optional method that updates the dataset by comparing against the remote. Not all unmanaged dataset backends
@@ -206,3 +208,95 @@ more, check out the docs here: [https://docs.gigantum.com](https://docs.gigantum
 
         # Run local update
         self.update_from_local(dataset, status_update_fn, status_result=status, verify_contents=True)
+
+    def update_from_local(self, dataset, status_update_fn: Callable,
+                          verify_contents: bool = False,
+                          status_result: Optional[StatusResult] = None) -> None:
+        """Method to update the dataset manifest for changed files that exists locally
+
+        Args:
+            dataset: Dataset object
+            status_update_fn: A callable, accepting a string for logging/providing status to the UI
+            verify_contents: Boolean indicating if "verify_contents" should be run, and the results added to modified
+            status_result: Optional StatusResult object to include in the update (typically from update_from_remote())
+
+        Returns:
+            None
+        """
+        if 'username' not in self.configuration:
+            raise ValueError("Dataset storage backend requires current logged in username to verify contents")
+        m = Manifest(dataset, self.configuration.get('username'))
+
+        status_update_fn("Updating Dataset manifest from local file state.")
+
+        if status_result is not None:
+            if status_result.modified is not None:
+                modified_keys = copy.deepcopy(status_result.modified)
+            else:
+                modified_keys = list()
+        else:
+            modified_keys = list()
+
+        if verify_contents:
+            modified_keys.extend(self.verify_contents(dataset, status_update_fn))
+
+        # Create StatusResult to force modifications
+        if status_result:
+            created_result = copy.deepcopy(status_result.created)
+            # Check if any directories got created
+            for key in status_result.created:
+                if key[-1] != '/':
+                    # a file
+                    if os.path.dirname(key) not in m.manifest:
+                        # Add the directory to the manifest
+                        created_result.append(f"{os.path.dirname(key)}/")
+
+            created_result = list(set(created_result))
+            if '/' in created_result:
+                created_result.remove('/')
+
+            # Combine a previous StatusResult object (typically from "update_from_remote")
+            status = StatusResult(created=created_result,
+                                  modified=modified_keys,
+                                  deleted=status_result.deleted)
+        else:
+            status = StatusResult(created=[], modified=modified_keys, deleted=[])
+
+        # Update the manifest
+        previous_revision = m.dataset_revision
+
+        m.update(status)
+        m.create_update_activity_record(status)
+
+        # Link the revision dir
+        m.link_revision()
+        if os.path.isdir(os.path.join(m.cache_mgr.cache_root, previous_revision)):
+            shutil.rmtree(os.path.join(m.cache_mgr.cache_root, previous_revision))
+
+        status_update_fn("Update complete.")
+
+
+class ExternalStorageBackend(LocalFilesystemBackend):
+    """Parent class extending LocalFilesystemBackend to include synchronization with a remote (such as S3)"""
+
+    @property
+    def can_update_from_remote(self) -> bool:
+        """Property indicating if this backend can automatically update its contents to the latest on the remote
+
+        Returns:
+            bool
+        """
+        return True
+
+    def update_from_remote(self, dataset, status_update_fn: Callable) -> None:
+        """Optional method that updates the dataset by comparing against the remote. Not all unmanaged dataset backends
+        will be able to do this.
+
+        Args:
+            dataset: Dataset object
+            status_update_fn: A callable, accepting a string for logging/providing status to the UI
+
+        Returns:
+            None
+        """
+        raise NotImplemented
